[rpc server]: extract rpc_module to utils.
This commit extracts the `rpc_module` from the servers to be shared.
It will help to re-use rpc modules within both the servers
