+++ utils/Cargo.toml
@@ -15,14 +15,27 @@ hyper14 = { package = "hyper", version = "0.14", default-features = false, featu
 jsonrpsee-types = { path = "../types", version = "0.2.0-alpha.6", optional = true }
 log = { version = "0.4", optional = true }
 rustc-hash = { version = "1", optional = true }
+rand = { version = "0.8", optional = true }
 serde = { version = "1.0", default-features = false, features = ["derive"], optional = true }
 serde_json = { version = "1", features = ["raw_value"], optional = true }
+parking_lot = { version = "0.11", optional = true }
 
 [features]
 default = []
 hyper_13 = ["hyper13", "futures-util", "jsonrpsee-types"]
 hyper_14 = ["hyper14", "futures-util", "jsonrpsee-types"]
-server = ["anyhow", "futures-channel", "futures-util", "jsonrpsee-types", "rustc-hash", "serde", "serde_json", "log"]
+server = [
+	"anyhow",
+	"futures-channel",
+	"futures-util",
+	"jsonrpsee-types",
+	"rustc-hash",
+	"serde",
+	"serde_json",
+	"log",
+	"parking_lot",
+	"rand"
+]
 
 [dev-dependencies]
 serde_json = "1.0"
