+++ utils/src/server/helpers.rs
@@ -1,24 +1,11 @@
-//! Shared helpers for JSON-RPC Servers.
+use crate::server::{RpcId, RpcSender};
 
 use futures_channel::mpsc;
 use futures_util::stream::StreamExt;
 use jsonrpsee_types::v2::error::{JsonRpcError, JsonRpcErrorCode, JsonRpcErrorObject};
-use jsonrpsee_types::v2::params::{RpcParams, TwoPointZero};
+use jsonrpsee_types::v2::params::TwoPointZero;
 use jsonrpsee_types::v2::response::JsonRpcResponse;
-use rustc_hash::FxHashMap;
 use serde::Serialize;
-use serde_json::value::RawValue;
-
-/// Connection ID.
-pub type ConnectionId = usize;
-/// Sender.
-pub type RpcSender<'a> = &'a mpsc::UnboundedSender<String>;
-/// RPC ID.
-pub type RpcId<'a> = Option<&'a RawValue>;
-/// Method registered in the server.
-pub type Method = Box<dyn Send + Sync + Fn(RpcId, RpcParams, RpcSender, ConnectionId) -> anyhow::Result<()>>;
-/// Methods registered in the Server.
-pub type Methods = FxHashMap<&'static str, Method>;
 
 /// Helper for sending JSON-RPC responses to the client
 pub fn send_response(id: RpcId, tx: RpcSender, result: impl Serialize) {
