+++ ws-server/Cargo.toml
@@ -16,8 +16,6 @@ futures-util = { version = "0.3", default-features = false, features = ["io"] }
 jsonrpsee-types = { path = "../types", version = "0.2.0-alpha.6" }
 jsonrpsee-utils = { path = "../utils", version = "0.2.0-alpha.6", features = ["server"] }
 log = "0.4"
-parking_lot = "0.11"
-rand = "0.8"
 rustc-hash = "1.1.0"
 serde = { version = "1", default-features = false, features = ["derive"] }
 serde_json = { version = "1", features = ["raw_value"] }
